+++ CHANGELOG.md
@@ -1,22 +1,33 @@
+## 1.1.0
+
+- Upgraded to Flutter 3.0.0 and fixed related issues.
+- Upgraded to reorderables: ^0.5.0 and fixed a bug with missing ScrollController
+
 ## 1.0.5
+
 - Bugfixes
 - Added dark mode support
 
 ## 1.0.4
+
 - copyright notice in the README file
 
 ## 1.0.3
+
 - fixed visual bug in stroop effect test
 - fixed bug in corsi bloc tapping test
 - updated example app
 
 ## 1.0.2
+
 - updated README.md
 - updated pubspec.yaml
 
 ## 1.0.1
+
 - updated README.md
 
 ## 1.0.0
+
 - initial release
 - Supports 14 cognitive tests